==> 3_Features_I_II_III_Combine.ipynb <==
import pandas as pd
import numpy as np
# Load the Drive helper and mount
from google.colab import drive
drive.mount('/content/drive')

simple_features = pd.read_csv('/content/drive/My Drive/semantic-question-matching/Features/features_I.csv', delimiter="\t")
distances = pd.read_csv('/content/drive/My Drive/semantic-question-matching/Features/features_III.csv')

simple_features.head(2)

distances.head(2)

distances.drop(['question1', 'question2', 'is_duplicate'], axis=1, inplace=True)

features_final = pd.concat([simple_features, distances], axis=1)

features_final.head(2)

features_final.to_csv('final_features.csv', index=False, sep='\t')
