==> 10_ArtificialNeuralNetwork.ipynb <==
import pandas as pd
import numpy as np

from keras.models import Sequential
from keras.layers import Dense, LSTM, Flatten, Embedding

from google.colab import drive
drive.mount('/content/drive')

dataset = pd.read_csv('/content/drive/My Drive/semantic-question-matching/Features/final_features.csv', delimiter='\t')
dataset.head(2)

dataset.shape

dataset.columns

dataset.replace(to_replace=np.inf, value=np.nan, inplace=True)
dataset.dropna(axis=0, inplace=True)

#diff_len, common_words,  fuzz_qratio ,fuzz_WRatio,fuzz_partial_ratio,fuzz_partial_token_set_ratio,  fuzz_partial_token_sort_ratio,  fuzz_token_set_ratio, fuzz_token_sort_ratio,cosine_distance, cityblock_distance, canberra_distance,euclidean_distance,minkowski_distance  and  braycurtis_distance. 
X = dataset.iloc[:, [5,10,11,12,13,14,15,16,17,18,20,21,23,24,25,26]].values
y = dataset.iloc[:, 2].values

np.any(np.isinf(X))

np.any(np.isnan(X))

from sklearn.model_selection import train_test_split
X_train, X_test, Y_train, Y_test = train_test_split(X, y, test_size=0.2)

print(np.sum(Y_train))
print(np.sum(Y_test))

from sklearn.preprocessing import StandardScaler
sc_X = StandardScaler()
X_train = sc_X.fit_transform(X_train)
X_test = sc_X.transform(X_test)

import pickle
pickle.dump(sc_X, open('XScaler', 'wb'))

classifier=Sequential()
classifier.add(Dense(input_dim=16, kernel_initializer="uniform", units=8, activation="relu"))
classifier.add(Dense(kernel_initializer="uniform", units=8, activation="relu"))
#classifier.add(Dense(kernel_initializer="uniform", units=8, activation="relu"))
#classifier.add(Dense(kernel_initializer="uniform", units=8, activation="relu"))
#classifier.add(Dense(kernel_initializer="uniform", units=8, activation="relu"))
#classifier.add(Dense(kernel_initializer="uniform", units=8, activation="relu"))
classifier.add(Dense(kernel_initializer="uniform", units=1, activation="sigmoid"))

classifier.compile(optimizer='adam',loss='binary_crossentropy',metrics=['accuracy'])

classifier.fit(X_train,Y_train,batch_size=10,epochs=50)

y_pred = classifier.predict(X_test)
y_pred=(y_pred>0.5)

from sklearn.metrics import confusion_matrix
cm = confusion_matrix(Y_test, y_pred)
print(cm)

import pickle
pickle.dump(classifier, open('ANN7274.model', 'wb'))
